allowed cogsLocation to be optional for normative

+++ Cogs.Publishers/UmlSchemaPublisher.cs
@@ -77,38 +77,41 @@ namespace Cogs.Publishers
             {
                 reusableList.Add(item.Name);
             }
-            // run svg publisher to create svg file to use for positioning 
-            SvgSchemaPublisher publisher = new SvgSchemaPublisher();
-            publisher.DotLocation = DotLocation;
-            publisher.TargetDirectory = TargetDirectory;
-            publisher.Overwrite = Overwrite;
-            publisher.Publish(model);
-
-            // read created svg file
-            XNamespace ns ="http://www.w3.org/2000/svg";
-            var nodes = XDocument.Load(Path.Combine(TargetDirectory, "output.svg")).Root.Descendants(ns +"g")
-                .Where(x => x.Attribute("class").Value == "node").ToList();
-            File.Delete(Path.Combine(TargetDirectory, "output.svg"));
-
-            //get leftmost  and topmost value to shift graph accordingly 
+            List<XElement> nodes = null;
             var xOff = 0.0;
             var yOff = 0.0;
-            foreach (var item in model.ItemTypes.Concat(model.ReusableDataTypes))
+            XNamespace ns = "http://www.w3.org/2000/svg";
+            if (!Normative)
             {
-                var node = XElement.Parse(nodes.Descendants(ns + "title").Where(x => x.FirstNode.ToString() == item.Name).ToList()[0].NextNode.ToString());
-                var cords = node.Attribute("points").Value.Split(',', ' ');
-                if(Convert.ToDouble(cords[4]) < xOff)
+                // run svg publisher to create svg file to use for positioning
+                SvgSchemaPublisher publisher = new SvgSchemaPublisher();
+                publisher.DotLocation = DotLocation;
+                publisher.TargetDirectory = TargetDirectory;
+                publisher.Overwrite = Overwrite;
+                publisher.Publish(model);
+
+                // read created svg file
+                nodes = XDocument.Load(Path.Combine(TargetDirectory, "output.svg")).Root.Descendants(ns + "g")
+                    .Where(x => x.Attribute("class").Value == "node").ToList();
+                File.Delete(Path.Combine(TargetDirectory, "output.svg"));
+
+                //get leftmost  and topmost value to shift graph accordingly 
+                foreach (var item in model.ItemTypes.Concat(model.ReusableDataTypes))
                 {
-                    xOff = Convert.ToDouble(cords[4]);
+                    var node = XElement.Parse(nodes.Descendants(ns + "title").Where(x => x.FirstNode.ToString() == item.Name).ToList()[0].NextNode.ToString());
+                    var cords = node.Attribute("points").Value.Split(',', ' ');
+                    if (Convert.ToDouble(cords[4]) < xOff)
+                    {
+                        xOff = Convert.ToDouble(cords[4]);
+                    }
+                    if (Convert.ToDouble(cords[3]) < yOff)
+                    {
+                        yOff = Convert.ToDouble(cords[3]);
+                    }
                 }
-                if(Convert.ToDouble(cords[3]) < yOff)
-                {
-                    yOff = Convert.ToDouble(cords[3]);                }
-                }
-            xOff = Math.Abs(xOff);
-            yOff = Math.Abs(yOff);
-            //get rightmost value
-         //   
+                xOff = Math.Abs(xOff);
+                yOff = Math.Abs(yOff);
+            } 
             int count = classList.Count;
             // loop through classes and reusable data types
             foreach (var item in model.ItemTypes.Concat(model.ReusableDataTypes))
@@ -118,15 +121,19 @@ namespace Cogs.Publishers
                            new XAttribute(xmins + "id", CreateId(item.Name)),
                            new XAttribute("name", item.Name)));
                 // add class to diagram
-                var cords = XElement.Parse(nodes.Descendants(ns + "title").Where(x => x.FirstNode.ToString() == item.Name).ToList()[0].NextNode.ToString())
+                if (!Normative)
+                {
+                    // add class to graph
+                    var cords = XElement.Parse(nodes.Descendants(ns + "title").Where(x => x.FirstNode.ToString() == item.Name).ToList()[0].NextNode.ToString())
                     .Attribute("points").Value.Split(',', ' ');
-                var left = (Convert.ToDouble(cords[4]) + xOff).ToString();
-                var right = (Convert.ToDouble(cords[0]) + xOff).ToString();
-                var top = (Convert.ToDouble(cords[1]) + yOff).ToString();
-                var bottom = (Convert.ToDouble(cords[3]) + yOff).ToString();
-                diagramElements.Add(new XElement("element", new XAttribute("geometry", "Left=" + left + ";Top=" + top + ";Right=" + right+ ";Bottom=" + bottom + ";"),
-                    new XAttribute("subject", item.Name), new XAttribute("seqno", count.ToString()), new XAttribute("style",
-                    "DUID=" + "item.Name" + ";NSL=0;BCol=-1;BFol=-1;LCol=-1;LWth=-1;fontsz=0;bold=0;black=0;italic=0;ul=0;charset=0;pitch=0;));")));
+                    var left = (Convert.ToDouble(cords[4]) + xOff).ToString();
+                    var right = (Convert.ToDouble(cords[0]) + xOff).ToString();
+                    var top = (Convert.ToDouble(cords[1]) + yOff).ToString();
+                    var bottom = (Convert.ToDouble(cords[3]) + yOff).ToString();
+                    diagramElements.Add(new XElement("element", new XAttribute("geometry", "Left=" + left + ";Top=" + top + ";Right=" + right + ";Bottom=" + bottom + ";"),
+                        new XAttribute("subject", item.Name), new XAttribute("seqno", count.ToString()), new XAttribute("style",
+                        "DUID=" + "item.Name" + ";NSL=0;BCol=-1;BFol=-1;LCol=-1;LWth=-1;fontsz=0;bold=0;black=0;italic=0;ul=0;charset=0;pitch=0;));")));
+                }
                 string extends = item.ExtendsTypeName;
                 // loop through properties of class and add to class
                 foreach(var property in item.Properties)
